Removed Builder class
Moved tests from ZosConnectionCreate to ZosConnectionFactory
