+++ src/main/java/zowe/client/sdk/core/ZosConnectionFactory.java
@@ -52,13 +52,10 @@ public class ZosConnectionFactory {
         if (password == null || password.isBlank())
             throw new IllegalStateException("password is either null or empty");
 
-        return new ZosConnection.Builder()
-                .host(host)
-                .zosmfPort(port)
-                .user(user)
-                .password(password)
-                .authType(AuthType.BASIC)
-                .build();
+        ZosConnection zosConnection = new ZosConnection(host, port, AuthType.BASIC);
+        zosConnection.setUser(user);
+        zosConnection.setPassword(password);
+        return zosConnection;
     }
 
     /**
@@ -82,12 +79,9 @@ public class ZosConnectionFactory {
         if (token == null)
             throw new IllegalStateException("token is null");
 
-        return new ZosConnection.Builder()
-                .host(host)
-                .zosmfPort(port)
-                .token(token)
-                .authType(AuthType.TOKEN)
-                .build();
+        ZosConnection zosConnection = new ZosConnection(host, port, AuthType.TOKEN);
+        zosConnection.setToken(token);
+        return zosConnection;
     }
 
     /**
@@ -116,13 +110,10 @@ public class ZosConnectionFactory {
         if (certPassword == null || certPassword.isBlank())
             throw new IllegalStateException("certificate password is either null or empty");
 
-        return new ZosConnection.Builder()
-                .host(host)
-                .zosmfPort(port)
-                .certFilePath(certFilePath)
-                .certPassword(certPassword)
-                .authType(AuthType.SSL)
-                .build();
+        ZosConnection zosConnection = new ZosConnection(host, port, AuthType.SSL);
+        zosConnection.setCertFilePath(certFilePath);
+        zosConnection.setCertPassword(certPassword);
+        return zosConnection;
     }
 
 }
